Velocity Control of a Block
I'm "controlling" a block sliding with friction by creating a velocity rpofile based on its position.

The real challenge for the quad is determining how to create a vel profile that has us slowing down as we approach something that is moving and not miss it.

Also, how to necessarily do this with LQR and the full state vector, but I don't think that's prohibitive -- we would set our qdes velocity terms to the reference.